fix: adding wrappers for general & address messages use cases
